feat: createLikeDao and FriendDao
