+++ src/main/java/ru/yandex/practicum/javafilmorate/service/FilmService.java
@@ -3,10 +3,7 @@ package ru.yandex.practicum.javafilmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.javafilmorate.storage.FilmDao;
-import ru.yandex.practicum.javafilmorate.storage.GenreDao;
-import ru.yandex.practicum.javafilmorate.storage.MpaDao;
-import ru.yandex.practicum.javafilmorate.storage.UserDao;
+import ru.yandex.practicum.javafilmorate.storage.*;
 import ru.yandex.practicum.javafilmorate.exception.NotFoundException;
 import ru.yandex.practicum.javafilmorate.model.Film;
 
@@ -21,6 +18,7 @@ public class FilmService {
     private final UserDao userStorage;
     private final MpaDao daoStorage;
     private final GenreDao genreStorage;
+    private final LikeDao likeStorage;
 
     public Film createFilm(Film film) {
         daoStorage.isMpaExisted(film.getMpa().getId());
@@ -52,7 +50,7 @@ public class FilmService {
     public void addLikes(int filmId, int userId) {
         filmStorage.isFilmExisted(filmId);
         userStorage.isUserExisted(userId);
-        filmStorage.createLike(filmId, userId);
+        likeStorage.createLike(filmId, userId);
         log.info("Film id: {} like from user: {} ", filmId, userId);
     }
 
@@ -61,7 +59,7 @@ public class FilmService {
             throw new NotFoundException("Negative value is not allowed");
         }
         Film film = getFilmById(filmId);
-        filmStorage.deleteLike(filmId, userId);
+        likeStorage.deleteLike(filmId, userId);
         log.info("The user with id = {} remove a like from the film id = {}", userId, filmId);
         return film;
     }
