+++ src/main/java/ru/yandex/practicum/javafilmorate/storage/LikeDao.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.javafilmorate.storage;
+
+public interface LikeDao {
+    void createLike(int id, int userId);
+
+    void deleteLike(int id, int userId);
+}
