+++ src/main/java/ru/yandex/practicum/javafilmorate/storage/UserDao.java
@@ -14,14 +14,6 @@ public interface UserDao {
 
     User updateUser(User user);
 
-    void addFriend(int id, int friendId);
-
-    void delete(int id, int friendId);
-
-    List<User> getCommonFriends(int id, int otherId);
-
-    List<User> getAllFriends(int id);
-
     void isUserExisted(int id);
 
 }
